Move rate limit methods to helper
