+++ test/integration/rack_attack_test.rb
@@ -1,100 +1,15 @@
 require "test_helper"
+require "helpers/rate_limit_helpers"
 
 class RackAttackTest < ActionDispatch::IntegrationTest
+  include RateLimitHelpers
+
   setup do
     Rack::Attack.cache.store = ActiveSupport::Cache::MemoryStore.new
     Rails.cache.clear
 
     @ip_address = "1.2.3.4"
     @user = create(:user, email: "nick@example.com", password: PasswordHelpers::SECURE_TEST_PASSWORD)
-  end
-
-  def exceeding_limit
-    (Rack::Attack::REQUEST_LIMIT * 1.25).to_i
-  end
-
-  def exceeding_email_limit
-    (Rack::Attack::REQUEST_LIMIT_PER_EMAIL * 1.25).to_i
-  end
-
-  def exceeding_exp_base_limit
-    (Rack::Attack::EXP_BASE_REQUEST_LIMIT * 1.25).to_i
-  end
-
-  def under_limit
-    (Rack::Attack::REQUEST_LIMIT * 0.5).to_i
-  end
-
-  def under_email_limit
-    (Rack::Attack::REQUEST_LIMIT_PER_EMAIL * 0.5).to_i
-  end
-
-  def limit_period
-    Rack::Attack::LIMIT_PERIOD
-  end
-
-  def push_limit_period
-    Rack::Attack::PUSH_LIMIT_PERIOD
-  end
-
-  def exp_base_limit_period
-    Rack::Attack::EXP_BASE_LIMIT_PERIOD
-  end
-
-  def exceed_limit_for(scope)
-    update_limit_for("#{scope}:#{@ip_address}", exceeding_limit)
-  end
-
-  def exceed_email_limit_for(scope)
-    update_limit_for("#{scope}:#{@user.email}", exceeding_email_limit)
-  end
-
-  def exceed_push_limit_for(scope)
-    exceeding_push_limit = (Rack::Attack::PUSH_LIMIT * 1.25).to_i
-    update_limit_for("#{scope}:#{@ip_address}", exceeding_push_limit, push_limit_period)
-  end
-
-  def exceed_exp_base_limit_for(scope)
-    update_limit_for("#{scope}:#{@ip_address}", exceeding_exp_base_limit, exp_base_limit_period)
-  end
-
-  def stay_under_limit_for(scope)
-    update_limit_for("#{scope}:#{@ip_address}", under_limit)
-  end
-
-  def stay_under_email_limit_for(scope)
-    update_limit_for("#{scope}:#{@user.email}", under_email_limit)
-  end
-
-  def stay_under_push_limit_for(scope)
-    under_push_limit = (Rack::Attack::PUSH_LIMIT * 0.5).to_i
-    update_limit_for("#{scope}:#{@user.email}", under_push_limit)
-  end
-
-  def stay_under_exp_base_limit_for(scope)
-    under_exp_base_limit = (Rack::Attack::EXP_BASE_REQUEST_LIMIT * 0.5).to_i
-    update_limit_for("#{scope}:#{@user.email}", under_exp_base_limit, exp_base_limit_period)
-  end
-
-  def update_limit_for(key, limit, period = limit_period)
-    limit.times { Rack::Attack.cache.count(key, period) }
-  end
-
-  def exceed_exponential_limit_for(scope, level)
-    expo_exceeding_limit = exceeding_exp_base_limit * level
-    expo_limit_period = exp_base_limit_period**level
-    expo_exceeding_limit.times { Rack::Attack.cache.count("#{scope}:#{@ip_address}", expo_limit_period) }
-  end
-
-  def encode(username, password)
-    ActionController::HttpAuthentication::Basic
-      .encode_credentials(username, password)
-  end
-
-  def expected_retry_after(level)
-    now = Time.now.to_i
-    period = Rack::Attack::EXP_BASE_LIMIT_PERIOD**level
-    (period - (now % period)).to_s
   end
 
   context "requests is lower than limit" do
@@ -149,45 +64,12 @@ class RackAttackTest < ActionDispatch::IntegrationTest
         @rubygem.ownerships.create(user: @user)
       end
 
-      should "allow gem yank by ip" do
-        stay_under_exp_base_limit_for("api/ip/1")
-
-        delete "/api/v1/gems/yank",
-          params: { gem_name: @rubygem.to_param, version: @rubygem.latest_version.number },
-          headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key }
-
-        assert_response :success
-      end
-
       should "allow gem push by ip" do
         stay_under_push_limit_for("api/push/ip")
 
         post "/api/v1/gems",
           params: gem_file("test-1.0.0.gem").read,
           headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key, CONTENT_TYPE: "application/octet-stream" }
-
-        assert_response :success
-      end
-
-      should "allow owner add by ip" do
-        second_user = create(:user)
-        stay_under_exp_base_limit_for("api/ip/1")
-
-        post "/api/v1/gems/#{@rubygem.name}/owners",
-          params: { rubygem_id: @rubygem.to_param, email: second_user.email },
-          headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key }
-
-        assert_response :success
-      end
-
-      should "allow owner remove by ip" do
-        second_user = create(:user)
-        @rubygem.ownerships.create(user: second_user)
-        stay_under_exp_base_limit_for("api/ip/1")
-
-        delete "/api/v1/gems/#{@rubygem.name}/owners",
-          params: { rubygem_id: @rubygem.to_param, email: second_user.email },
-          headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key }
 
         assert_response :success
       end
@@ -239,6 +121,71 @@ class RackAttackTest < ActionDispatch::IntegrationTest
             headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key, CONTENT_TYPE: "application/octet-stream" }
 
           assert_response :ok
+        end
+      end
+
+      context "ui requests" do
+        setup do
+          @user.enable_mfa!(ROTP::Base32.random_base32, :ui_only)
+          stay_under_exponential_limit("clearance/ip")
+        end
+
+        should "allow for mfa sign in" do
+          post "/session", params: { session: { who: @user.handle, password: @user.password } } # sets session[:mfa_user]
+
+          post "/session/mfa_create",
+            params: { otp: ROTP::TOTP.new(@user.mfa_seed).now },
+            headers: { REMOTE_ADDR: @ip_address }
+
+          assert_redirected_to "/dashboard"
+        end
+
+        should "allow mfa forgot password" do
+          @user.forgot_password!
+          post "/users/#{@user.id}/password/mfa_edit",
+            params: { token: @user.confirmation_token, otp: ROTP::TOTP.new(@user.mfa_seed).now },
+            headers: { REMOTE_ADDR: @ip_address }
+
+          assert_response :ok
+        end
+      end
+
+      context "api requests" do
+        setup do
+          @user.enable_mfa!(ROTP::Base32.random_base32, :ui_and_api)
+          stay_under_exponential_limit("api/ip")
+
+          @rubygem = create(:rubygem, name: "test", number: "0.0.1")
+          @rubygem.ownerships.create(user: @user)
+        end
+
+        should "allow gem yank by ip" do
+          delete "/api/v1/gems/yank",
+            params: { gem_name: @rubygem.to_param, version: @rubygem.latest_version.number },
+            headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key, HTTP_OTP: ROTP::TOTP.new(@user.mfa_seed).now }
+
+          assert_response :success
+        end
+
+        should "allow owner add by ip" do
+          second_user = create(:user)
+
+          post "/api/v1/gems/#{@rubygem.name}/owners",
+            params: { rubygem_id: @rubygem.to_param, email: second_user.email },
+            headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key, HTTP_OTP: ROTP::TOTP.new(@user.mfa_seed).now }
+
+          assert_response :success
+        end
+
+        should "allow owner remove by ip" do
+          second_user = create(:user)
+          @rubygem.ownerships.create(user: second_user)
+
+          delete "/api/v1/gems/#{@rubygem.name}/owners",
+            params: { rubygem_id: @rubygem.to_param, email: second_user.email },
+            headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key, HTTP_OTP: ROTP::TOTP.new(@user.mfa_seed).now }
+
+          assert_response :success
         end
       end
     end
@@ -339,16 +286,6 @@ class RackAttackTest < ActionDispatch::IntegrationTest
         @rubygem.ownerships.create(user: @user)
       end
 
-      should "throttle gem yank by ip" do
-        exceed_exp_base_limit_for("api/ip/1")
-
-        delete "/api/v1/gems/yank",
-          params: { gem_name: @rubygem.to_param, version: @rubygem.latest_version.number },
-          headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key }
-
-        assert_response :too_many_requests
-      end
-
       should "throttle gem push by ip" do
         exceed_push_limit_for("api/push/ip")
 
@@ -358,42 +295,18 @@ class RackAttackTest < ActionDispatch::IntegrationTest
 
         assert_response :too_many_requests
       end
-
-      should "throttle owner add by ip" do
-        second_user = create(:user)
-        exceed_exp_base_limit_for("api/ip/1")
-
-        post "/api/v1/gems/#{@rubygem.name}/owners",
-          params: { rubygem_id: @rubygem.to_param, email: second_user.email },
-          headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key }
-
-        assert_response :too_many_requests
-      end
-
-      should "throttle owner remove by ip" do
-        second_user = create(:user)
-        @rubygem.ownerships.create(user: second_user)
-        exceed_exp_base_limit_for("api/ip/1")
-
-        delete "/api/v1/gems/#{@rubygem.name}/owners",
-          params: { rubygem_id: @rubygem.to_param, email: second_user.email },
-          headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key }
-
-        assert_response :too_many_requests
-      end
     end
 
     context "exponential backoff" do
+      setup { @mfa_max_period = { 1 => 100, 2 => 10_000, 3 => 1_000_000 } }
+
       Rack::Attack::EXP_BACKOFF_LEVELS.each do |level|
         should "throttle for mfa sign in at level #{level}" do
           freeze_time do
             exceed_exponential_limit_for("clearance/ip/#{level}", level)
+            post "/session/mfa_create", headers: { REMOTE_ADDR: @ip_address }
 
-            post "/session/mfa_create",
-              headers: { REMOTE_ADDR: @ip_address }
-
-            assert_response :too_many_requests
-            assert_equal expected_retry_after(level), @response.headers["Retry-After"]
+            assert_throttle_at(level)
           end
         end
 
@@ -405,53 +318,70 @@ class RackAttackTest < ActionDispatch::IntegrationTest
               params: gem_file("test-0.0.0.gem").read,
               headers: { REMOTE_ADDR: @ip_address, HTTP_AUTHORIZATION: @user.api_key, CONTENT_TYPE: "application/octet-stream" }
 
-            assert_response :too_many_requests
-            assert_equal expected_retry_after(level), @response.headers["Retry-After"]
+            assert_throttle_at(level)
           end
         end
 
         should "throttle mfa create at level #{level}" do
           freeze_time do
             exceed_exponential_limit_for("clearance/ip/#{level}", level)
+            post "/multifactor_auth", headers: { REMOTE_ADDR: @ip_address }
 
-            post "/multifactor_auth",
-              headers: { REMOTE_ADDR: @ip_address }
-            assert_response :too_many_requests
-            assert_equal expected_retry_after(level), @response.headers["Retry-After"]
+            assert_throttle_at(level)
           end
         end
 
         should "throttle mfa update at level #{level}" do
           freeze_time do
             exceed_exponential_limit_for("clearance/ip/#{level}", level)
+            put "/multifactor_auth", headers: { REMOTE_ADDR: @ip_address }
 
-            put "/multifactor_auth",
-              headers: { REMOTE_ADDR: @ip_address }
-            assert_response :too_many_requests
-            assert_equal expected_retry_after(level), @response.headers["Retry-After"]
+            assert_throttle_at(level)
           end
         end
 
         should "throttle api key show at level #{level}" do
           freeze_time do
             exceed_exponential_limit_for("api/ip/#{level}", level)
+            get "/api/v1/api_key.json", headers: { REMOTE_ADDR: @ip_address }
 
-            get "/api/v1/api_key.json",
-              headers: { REMOTE_ADDR: @ip_address }
-            assert_response :too_many_requests
-            assert_equal expected_retry_after(level), @response.headers["Retry-After"]
+            assert_throttle_at(level)
           end
         end
 
         should "throttle mfa forgot password at level #{level}" do
           freeze_time do
             exceed_exponential_limit_for("clearance/ip/#{level}", level)
+            post "/users/#{@user.id}/password/mfa_edit", headers: { REMOTE_ADDR: @ip_address }
 
-            post "/users/#{@user.id}/password/mfa_edit",
-              headers: { REMOTE_ADDR: @ip_address }
+            assert_throttle_at(level)
+          end
+        end
 
-            assert_response :too_many_requests
-            assert_equal expected_retry_after(level), @response.headers["Retry-After"]
+        should "throttle gem yank by ip #{level}" do
+          freeze_time do
+            exceed_exponential_limit_for("api/ip/#{level}", level)
+            delete "/api/v1/gems/yank", headers: { REMOTE_ADDR: @ip_address }
+
+            assert_throttle_at(level)
+          end
+        end
+
+        should "throttle owner add by ip #{level}" do
+          freeze_time do
+            exceed_exponential_limit_for("api/ip/#{level}", level)
+            post "/api/v1/gems/somegem/owners", headers: { REMOTE_ADDR: @ip_address }
+
+            assert_throttle_at(level)
+          end
+        end
+
+        should "throttle owner remove by ip #{level}" do
+          freeze_time do
+            exceed_exponential_limit_for("api/ip/#{level}", level)
+            delete "/api/v1/gems/somegem/owners", headers: { REMOTE_ADDR: @ip_address }
+
+            assert_throttle_at(level)
           end
         end
       end
